add bordeaux realtime
for the next departure I'm a bit confused.

There is no link that I can find on opendatasoft.

So the "best" thing I can find is to directly give the url of the
webservice, but so there are no explanation.

They have a demo site for their webservices, but I cannot give a precise
url for this...