add a link to the /datasets/{id} api in each dataset page
This way make a bit of advertising to our api.

Note that here I put a direct link to the api whereas in the datasets
index the link point to the `/datasets` section of the swagger ui.

I don't think we can poin to `/datasets/{id} in the swagger ui and since
there are no parameter to this api, I thought it was enough to add a
direct link.

This can be improved when we'll have a `API` page in the web site with
some explanations.