cleanup static files
remove the global route `/<file_name>`.

This was making 500 error when the file was not found (instead of 404)
and was cluttering the logs.

In the mean time I removed the static file `faq`, `guide`, and `legal`
that seems to be deprecated and redirect them on
doc.transport.data.gouv.fr