replace the dataset_geographic_view by a materialized view
The old view was causing some performance problem since we did a lot on
join on it (and it was dumb since the view has to be create at each time
it is queried).
This was causing a quadratic query for example on `/stats/regions` and
some timeouts.

Replacing the view by a materialized view fixes all the performance
issues