use a materialized view for search
use a materialized view for autocomplete instead of multiple unions

This makes it easier to tweak search and more performant mainly by
adding a composite field(indexed_name), cleaned up and indexed for
search.

This makes it possible to search for "ile de france" or "i d f".

We loose some control though, we cannot limit by type not (like asking
for 5 cities, 3 aoms, and 2 regions).

dumb bench or queries using the median of an `ab -n 10`.
(knowing that a huge part is not due to ecto, but it's easier to use
`ab`
than to check the db query time to get some rough estimation)

| search        | before                | after  |
| ------------- | --------------------- | ------ |
| montreuil     | 130 ms                | 40 ms  |
| pa            | 110 ms                | 92 ms  |
| a             | 129 ms                | 150 ms |
| ile de france | 1 ms (but no results) | 1 ms   |
| ile-de-france | 120 ms                | 1 ms   |