add a daily task to generate a NeTEx for each GTFS
This NeTEx will be generated through
https://github.com/etalab/gtfs_converter/ using kisio's
[gtfs2netexfr](https://github.com/CanalTP/transit_model/tree/master/gtfs2netexfr).

The NeTEx will be posted asynchronously on data.gouv as a commnunautary
resource.

We store the content_hash of the gtfs that has been used to generate the
netex not to regenerate a new resource needlessly (even if we try to
rewrite the same communautary resource when the GTFS changes).