Do not preload all resources / validation while revalidating all
The import/validate process seems to consume too much memory.

This reduce the memory footprint by not loading in memory all resources
and their validation.

With a dump test on my computer the `maxresident` of `/usr/bin/time`
says that the validate process is now 397_428 compared to 1_836_576
before.

I also changed the time to 4am, this way no user should be impacted if
some stuff slow down