Mise en cache des données coûteuses (home et stats) (#1530)
* Move all the SQL-generating stuff to a specific function

This will make caching the content easier.

* Start a Cachex cache for transport

* Implement a degradable 60-second cache for costly home page data

* Lower the ecto pool size on "prochainement" (#1528)

I am attempting to deploy the branch for testing.

* Fix typo

* Make credo happy

* Update prod.exs

* Extract caching logic to a specific module

Co-authored-by: Francis Chabouis <fchabouis@gmail.com>

* Rename method for clarity

* Enable caching for stats page main action

Co-authored-by: Francis Chabouis <fchabouis@gmail.com>

* Support optional caching for rendered features

* Enable cache for region stats

This lowers the computation from 250 to 50 ms. We could do better by converting the Elixir structure to a binary (to be done later).

* Cache other costly action

* Cache quality costly action

* Add global middleware to avoid encoding already JSON-encoded binaries

* Improve documentation

* Optimize caching speed x100 by avoiding marshalling & re-encoding to JSON

* Remove reference to avoid confusion

2 format_encoders for a given format won't work.

* Use an explicit tuple

If we ever render a single string as JSON response, the previous code would not have worked, because it would have not encoded the JSON string response (assuming it was already encoded).

Instead, this explicit tuple makes sure we have an explicit contract between the caller and the middleware.

* Add logging to help diagnose future issues

The whole logic can be a bit surprising at first, so this will be helpful to future maintainers.

* Improve documentation

* Rename component for clarity

* Add documentation

* Add documentation

* Revert now useless change

* Remove warning caused by @doc with a defp method

* Extract caching behaviour

This will allow to use a non-cache during tests, except where specified.

* Specify cache implementations

* Replace caching calls with new central API

* Add mox for more strict behaviour-based mocking

* Fix credo warnings

* Avoid the need to swap the impl via App config at test runtime

We use the Mox mock in the configuration, but extend it with interesting default behaviour (= no cache). This means most tests are not impacted, but we can still use Mox expectations, without swapping the App config.

* Assert that something is cached on the home

* Stop using Mox for now (failures with integration tests)

I'll revert to a "null cache" and add a few expectations with a different system for now.

* Remove focus tag

* Avoid raising on unexpected cases (e.g. :ignore) and prefer fallback

* Add some tests for the Cachex wrapper

* Add a first test to cover one stats API route with caching

* Comply with credo

* Add tests for all cached stats routes

* Reduce changes in this specific PR by reverting mix cleanup

This is taken from "master".

* Force deploy [skip ci]

* Force CI build

* Add placeholders for tests to be implemented

* Rename late-evaluated queries for clarity

We want to underline the fact that the evaluation won't be done right away.

* Go async for tests

The tests use UUID for cache keys, globally unique so no conflict will happen.

* Add a bit of code to detect potential future changes in Cachex output

The API is not super clearly defined.

* Remove covered case

* Catch computation exceptions and wrap them up with a specific tuple

* Make it easy to override cache name for testing cache errors

* Re-raise computation exceptions

Since we have a clearly defined tuple for computation errors, we can differentiate between Cachex-related errors and the ones which are raised by the caller code.

* Mention :ignore mode (not used currently)

* Finalize tests

* Add computation

* Add note

Co-authored-by: Francis Chabouis <fchabouis@gmail.com>