Make parent app use cached config by default
`config.exs` currently points to `fetch_config!/0`, so this changes means that the config will be cached by default (which is what we want for long running stuff).

A side effect that I wanted is that the backoffice proxy config view will then use exactly the same copy for display, hence showing a correct view of what is actually used for processing.