Avoid compile-time config evaluation
Moving to runtime allows us to work with Elixir releases later.

I'm not filtering on web-server only because the place where the plug is used is already web-server specific (the worker part is stopped earlier in the stack).

I'm not filtering on Mix.env either because we want to be able to verify the behaviour in dev (using `/etc/hosts`) and test, to later fix the numerous broken tests.